moved signup logic to login. there is no need for 2 pages
